cleanup

+++ src/BuildOrderSearchGoal.h
@@ -1,3 +1,5 @@
+/* -*- c-basic-offset: 4 -*- */
+
 #pragma once
 
 #include "Common.h"
@@ -6,28 +8,28 @@
 
 namespace BOSS
 {
-class BuildOrderSearchGoal
-{
-    std::vector<size_t> m_goalUnits;                 // vector of goal number of units indexed by ActionType ID
-    std::vector<size_t> m_goalUnitsMax;              // vector of goal max number of units indexed by ActionType ID
-    size_t              m_supplyRequiredVal;         // amount of supply required for all goal units in _goalUnits
+    class BuildOrderSearchGoal
+    {
+        std::vector<size_t> m_goalUnits;                 // vector of goal number of units indexed by ActionType ID
+        std::vector<size_t> m_goalUnitsMax;              // vector of goal max number of units indexed by ActionType ID
+        size_t              m_supplyRequiredVal;         // amount of supply required for all goal units in _goalUnits
 
-    void calculateSupplyRequired();
+        void calculateSupplyRequired();
 
-public:
+    public:
 
-    BuildOrderSearchGoal();
+        BuildOrderSearchGoal();
 
-    bool                operator == (const BuildOrderSearchGoal & g);
-    bool                hasGoal() const;
-    bool                isAchievedBy(const GameState & state);
+        bool                operator == (const BuildOrderSearchGoal & g);
+        bool                hasGoal() const;
+        bool                isAchievedBy(const GameState & state);
 
-    size_t              supplyRequired() const;
-    size_t              getGoal(ActionType a) const;
-    size_t              getGoalMax(ActionType a) const;
+        int                 supplyRequired() const;
+        int                 getGoal(ActionType a) const;
+        int                 getGoalMax(ActionType a) const;
 
-    void                setGoal(ActionType a, const size_t num);
-    void                setGoalMax(ActionType a, const size_t num);
-    std::string         toString() const;
-};
+        void                setGoal(ActionType a, const size_t num);
+        void                setGoalMax(ActionType a, const size_t num);
+        std::string         toString() const;
+    };
 }
