+++ src/CombatSearch_IntegralDataFinishedUnits.h
@@ -1,3 +1,5 @@
+/* -*- c-basic-offset: 4 -*- */
+
 #pragma once
 
 #include "Common.h"
@@ -13,14 +15,14 @@ namespace BOSS
     class IntegralDataFinishedUnits
     {
     public:
-        double  eval;
-        double  integral_ToThisPoint;
-        double  integral_UntilFrameLimit;
-        size_t     timeStarted;
-        size_t     timeFinished;
-        char id;
+        float   eval;
+        float   integral_ToThisPoint;
+        float   integral_UntilFrameLimit;
+        int     timeStarted;
+        int     timeFinished;
+        char    id;
 
-        IntegralDataFinishedUnits(double e, double i_c, double i_l, size_t t_a, size_t t_f)
+        IntegralDataFinishedUnits(float e, float i_c, float i_l, int t_a, int t_f)
             : eval(e)
             , integral_ToThisPoint(i_c)
             , integral_UntilFrameLimit(i_l)
@@ -31,7 +33,7 @@ namespace BOSS
 
         }
 
-        IntegralDataFinishedUnits(double e, double i_c, double i_l, size_t t_a, size_t t_f, char newID)
+        IntegralDataFinishedUnits(float e, float i_c, float i_l, int t_a, int t_f, char newID)
             : eval(e)
             , integral_ToThisPoint(i_c)
             , integral_UntilFrameLimit(i_l)
@@ -57,12 +59,12 @@ namespace BOSS
     class CombatSearch_IntegralDataFinishedUnits
     {
         std::vector<IntegralDataFinishedUnits>       m_integralStack;
-        size_t                                       m_chronoBoostEntries;
+        int m_chronoBoostEntries;
 
-        std::vector<IntegralDataFinishedUnits>       m_bestIntegralStack;
-        double                                       m_bestIntegralValue;
-        BuildOrderAbilities                          m_bestIntegralBuildOrder;
-        GameState                                    m_bestIntegralGameState;
+        std::vector<IntegralDataFinishedUnits>      m_bestIntegralStack;
+        float                                       m_bestIntegralValue;
+        BuildOrderAbilities                         m_bestIntegralBuildOrder;
+        GameState                                   m_bestIntegralGameState;
 
         BuildOrderAbilities createBuildOrderEndTimes(const std::vector<IntegralDataFinishedUnits> & integral_stack, const GameState & state) const;
 
@@ -75,10 +77,9 @@ namespace BOSS
         void popFinishedLastOrder(const GameState & prevState, const GameState & currState);
 
         void print() const;
-        void printIntegralData(const size_t index, const std::vector<IntegralDataFinishedUnits> & integral_stack, const GameState & state, const BuildOrderAbilities & buildOrder) const;
+        void printIntegralData(const int index, const std::vector<IntegralDataFinishedUnits> & integral_stack, const GameState & state, const BuildOrderAbilities & buildOrder) const;
         void print(const std::vector<IntegralDataFinishedUnits> & integral_stack, const GameState & state, const BuildOrderAbilities & buildOrder) const;
 
         const BuildOrderAbilities & getBestBuildOrder() const;
     };
-
 }
